+++ stock-common/src/main/java/co/handk/common/model/dto/query/MessageQueryDTO.java
@@ -11,7 +11,7 @@ public class MessageQueryDTO extends PageQuery {
     private Integer type;
     private Long userId;
     private String message;
-    private Long sourceId;
+    private Integer sourceId;
     private Integer isRead;
     private Integer state;
 }
