+++ stock-common/src/main/java/co/handk/common/model/vo/MessageVO.java
@@ -7,7 +7,7 @@ public class MessageVO extends BaseVO {
     private Integer type;
     private Long userId;
     private String message;
-    private Long sourceId;
+    private Integer sourceId;
     private Integer isRead;
     private Integer state;
 }
